Added unit test for jetstream utils
