+++ components/eventing-controller/pkg/backend/jetstreamv2/jetstream_test.go
@@ -2,6 +2,11 @@ package jetstreamv2
 
 import (
 	"fmt"
+	"net"
+	"strings"
+	"testing"
+	"time"
+
 	kymalogger "github.com/kyma-project/kyma/common/logging/logger"
 	eventingv1alpha2 "github.com/kyma-project/kyma/components/eventing-controller/api/v1alpha2"
 	"github.com/kyma-project/kyma/components/eventing-controller/logger"
@@ -15,22 +20,12 @@ import (
 	"github.com/nats-io/nats-server/v2/server"
 	"github.com/nats-io/nats.go"
 	"github.com/pkg/errors"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"net"
-	"reflect"
-	"strings"
-	"testing"
-	"time"
 )
 
 const (
 	defaultStreamName    = "kyma"
 	defaultMaxReconnects = 10
-
-	// maxJetStreamConsumerNameLength is the maximum preferred length for the JetStream consumer names
-	// as per https://docs.nats.io/running-a-nats-service/nats_admin/jetstream_admin/naming
-	maxJetStreamConsumerNameLength = 32
 )
 
 // TestJetStreamInitialize_NoStreamExists tests if a stream is
@@ -712,67 +707,6 @@ func TestJetStreamSubAfterSync_MultipleSubs(t *testing.T) {
 	require.Equal(t, subBytesLimit, bytesLimit)
 }
 
-// TestJetStream_isJsSubAssociatedWithKymaSub tests the isJsSubAssociatedWithKymaSub method.
-func TestJetStream_isJsSubAssociatedWithKymaSub(t *testing.T) {
-	// given
-	testEnvironment := setupTestEnvironment(t)
-	jsBackend := testEnvironment.jsBackend
-	defer testEnvironment.natsServer.Shutdown()
-	defer testEnvironment.jsClient.natsConn.Close()
-	initErr := jsBackend.Initialize(nil)
-	require.NoError(t, initErr)
-
-	// create subscription 1 and its JetStream subscription
-	cleanSubject1 := "subOne"
-	sub1 := evtestingv2.NewSubscription(cleanSubject1, "foo", evtestingv2.WithNotCleanEventSourceAndType())
-	jsSub1Key := NewSubscriptionSubjectIdentifier(sub1, cleanSubject1)
-
-	// create subscription 2 and its JetStream subscription
-	cleanSubject2 := "subOneTwo"
-	sub2 := evtestingv2.NewSubscription(cleanSubject2, "foo", evtestingv2.WithNotCleanEventSourceAndType())
-	jsSub2Key := NewSubscriptionSubjectIdentifier(sub2, cleanSubject2)
-
-	testCases := []struct {
-		name            string
-		givenJSSubKey   SubscriptionSubjectIdentifier
-		givenKymaSubKey *eventingv1alpha2.Subscription
-		wantResult      bool
-	}{
-		{
-			name:            "",
-			givenJSSubKey:   jsSub1Key,
-			givenKymaSubKey: sub1,
-			wantResult:      true,
-		},
-		{
-			name:            "",
-			givenJSSubKey:   jsSub2Key,
-			givenKymaSubKey: sub2,
-			wantResult:      true,
-		},
-		{
-			name:            "",
-			givenJSSubKey:   jsSub1Key,
-			givenKymaSubKey: sub2,
-			wantResult:      false,
-		},
-		{
-			name:            "",
-			givenJSSubKey:   jsSub2Key,
-			givenKymaSubKey: sub1,
-			wantResult:      false,
-		},
-	}
-
-	for _, tC := range testCases {
-		testCase := tC
-		t.Run(testCase.name, func(t *testing.T) {
-			gotResult := isJsSubAssociatedWithKymaSub(tC.givenJSSubKey, tC.givenKymaSubKey)
-			require.Equal(t, tC.wantResult, gotResult)
-		})
-	}
-}
-
 // TestMultipleJSSubscriptionsToSameEvent tests the behaviour of JS
 // when multiple subscriptions need to receive the same event.
 func TestMultipleJSSubscriptionsToSameEvent(t *testing.T) {
@@ -1216,247 +1150,6 @@ func TestJetStream_ServerAndSinkRestart(t *testing.T) {
 	require.NoError(t, newSubscriber.CheckEvent(expectedEv2Data))
 }
 
-func defaultNatsConfig(url string) env.NatsConfig {
-	return env.NatsConfig{
-		URL:                     url,
-		MaxReconnects:           defaultMaxReconnects,
-		ReconnectWait:           3 * time.Second,
-		JSStreamName:            defaultStreamName,
-		JSStreamStorageType:     StorageTypeMemory,
-		JSStreamRetentionPolicy: RetentionPolicyInterest,
-	}
-}
-
-// getJetStreamClient creates a client with JetStream context, or fails the caller test.
-func getJetStreamClient(t *testing.T, serverURL string) *jetStreamClient {
-	conn, err := nats.Connect(serverURL)
-	if err != nil {
-		t.Error(err.Error())
-	}
-	jsCtx, err := conn.JetStream()
-	if err != nil {
-		conn.Close()
-		t.Error(err.Error())
-	}
-	return &jetStreamClient{
-		JetStreamContext: jsCtx,
-		natsConn:         conn,
-	}
-}
-
-func addJSCleanEventTypesToStatus(sub *eventingv1alpha2.Subscription, cleaner cleaner.Cleaner) error {
-	cleanEventType, err := getCleanEventTypes(sub, cleaner)
-	if err != nil {
-		return err
-	}
-	sub.Status.Types = cleanEventType
-	return nil
-}
-
-// TestEnvironment provides mocked resources for tests.
-type TestEnvironment struct {
-	jsBackend  *JetStream
-	logger     *logger.Logger
-	natsServer *server.Server
-	jsClient   *jetStreamClient
-	natsConfig env.NatsConfig
-	cleaner    cleaner.Cleaner
-	natsPort   int
-}
-
-// setupTestEnvironment is a TestEnvironment constructor.
-func setupTestEnvironment(t *testing.T) *TestEnvironment {
-	natsServer, natsPort, err := natstesting.StartNATSServer(evtesting.WithJetStreamEnabled())
-	require.NoError(t, err)
-	natsConfig := defaultNatsConfig(natsServer.ClientURL())
-	defaultLogger, err := logger.New(string(kymalogger.JSON), string(kymalogger.INFO))
-	require.NoError(t, err)
-
-	// init the metrics collector
-	metricsCollector := metrics.NewCollector()
-
-	jsClient := getJetStreamClient(t, natsConfig.URL)
-	jsCleaner := cleaner.NewJetStreamCleaner(defaultLogger)
-	jsBackend := NewJetStream(natsConfig, metricsCollector, jsCleaner, defaultLogger)
-
-	return &TestEnvironment{
-		jsBackend:  jsBackend,
-		logger:     defaultLogger,
-		natsServer: natsServer,
-		jsClient:   jsClient,
-		natsConfig: natsConfig,
-		cleaner:    jsCleaner,
-		natsPort:   natsPort,
-	}
-}
-
-// TestSubscriptionSubjectIdentifierEqual checks the equality of two SubscriptionSubjectIdentifier instances and their consumer names.
-func TestSubscriptionSubjectIdentifierEqual(t *testing.T) {
-	t.Parallel()
-	testCases := []struct {
-		name             string
-		givenIdentifier1 SubscriptionSubjectIdentifier
-		givenIdentifier2 SubscriptionSubjectIdentifier
-		wantEqual        bool
-	}{
-		// instances are equal
-		{
-			name: "should be equal if the two instances are identical",
-			givenIdentifier1: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-1", "ns-1"),
-				"prefix.app.event.operation.v1",
-			),
-			givenIdentifier2: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-1", "ns-1"),
-				"prefix.app.event.operation.v1",
-			),
-			wantEqual: true,
-		},
-		// instances are not equal
-		{
-			name: "should not be equal if only name is different",
-			givenIdentifier1: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-1", "ns-1"),
-				"prefix.app.event.operation.v1",
-			),
-			givenIdentifier2: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-2", "ns-1"),
-				"prefix.app.event.operation.v1",
-			),
-			wantEqual: false,
-		},
-		{
-			name: "should not be equal if only namespace is different",
-			givenIdentifier1: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-1", "ns-1"),
-				"prefix.app.event.operation.v1",
-			),
-			givenIdentifier2: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-1", "ns-2"),
-				"prefix.app.event.operation.v1",
-			),
-			wantEqual: false,
-		},
-		{
-			name: "should not be equal if only subject is different",
-			givenIdentifier1: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-1", "ns-1"),
-				"prefix.app.event.operation.v1",
-			),
-			givenIdentifier2: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-1", "ns-1"),
-				"prefix.app.event.operation.v2",
-			),
-			wantEqual: false,
-		},
-		// possible naming collisions
-		{
-			name: "should not be equal if subject is the same but name and namespace are swapped",
-			givenIdentifier1: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-1", "ns-1"),
-				"prefix.app.event.operation.v1",
-			),
-			givenIdentifier2: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("ns-1", "sub-1"),
-				"prefix.app.event.operation.v1",
-			),
-			wantEqual: false,
-		},
-		{
-			name: "should not be equal if subject is the same but name and namespace are only equal if joined together",
-			givenIdentifier1: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-1", "ns-1"), // evaluates to "sub-1ns-1" when joined
-				"prefix.app.event.operation.v1",
-			),
-			givenIdentifier2: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub-1n", "s-1"), // evaluates to "sub-1ns-1" when joined
-				"prefix.app.event.operation.v1",
-			),
-			wantEqual: false,
-		},
-	}
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			gotInstanceEqual := reflect.DeepEqual(tc.givenIdentifier1, tc.givenIdentifier2)
-			assert.Equal(t, tc.wantEqual, gotInstanceEqual)
-
-			gotConsumerNameEqual := tc.givenIdentifier1.ConsumerName() == tc.givenIdentifier2.ConsumerName()
-			assert.Equal(t, tc.wantEqual, gotConsumerNameEqual)
-		})
-	}
-}
-
-// TestSubscriptionSubjectIdentifierConsumerNameLength checks that the SubscriptionSubjectIdentifier consumer name
-// length is equal to the recommended length by JetStream.
-func TestSubscriptionSubjectIdentifierConsumerNameLength(t *testing.T) {
-	t.Parallel()
-	testCases := []struct {
-		name                   string
-		givenIdentifier        SubscriptionSubjectIdentifier
-		wantConsumerNameLength int
-	}{
-		{
-			name: "short string values",
-			givenIdentifier: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub", "ns"),
-				"app.event.operation.v1",
-			),
-			wantConsumerNameLength: maxJetStreamConsumerNameLength,
-		},
-		{
-			name: "long string values",
-			givenIdentifier: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("some-test-subscription", "some-test-namespace"),
-				"some.test.prefix.some-test-application.some-test-event-name.some-test-operation.some-test-version",
-			),
-			wantConsumerNameLength: maxJetStreamConsumerNameLength,
-		},
-	}
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			assert.Equal(t, tc.wantConsumerNameLength, len(tc.givenIdentifier.ConsumerName()))
-		})
-	}
-}
-
-// TestSubscriptionSubjectIdentifierNamespacedName checks the syntax of the SubscriptionSubjectIdentifier namespaced name.
-func TestSubscriptionSubjectIdentifierNamespacedName(t *testing.T) {
-	t.Parallel()
-	testCases := []struct {
-		name               string
-		givenIdentifier    SubscriptionSubjectIdentifier
-		wantNamespacedName string
-	}{
-		{
-			name: "short name and namespace values",
-			givenIdentifier: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("sub", "ns"),
-				"app.event.operation.v1",
-			),
-			wantNamespacedName: "ns/sub",
-		},
-		{
-			name: "long name and namespace values",
-			givenIdentifier: NewSubscriptionSubjectIdentifier(
-				evtestingv2.NewSubscription("some-test-subscription", "some-test-namespace"),
-				"app.event.operation.v1",
-			),
-			wantNamespacedName: "some-test-namespace/some-test-subscription",
-		},
-	}
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			assert.Equal(t, tc.wantNamespacedName, tc.givenIdentifier.NamespacedName())
-		})
-	}
-}
-
 // TestJetStream_NoNATSSubscription tests if the error is being triggered
 // when expected entries in js.subscriptions map are missing.
 func TestJetStream_NATSSubscriptionCount(t *testing.T) {
@@ -1530,7 +1223,7 @@ func TestJetStream_NATSSubscriptionCount(t *testing.T) {
 			// when
 			err := jsBackend.SyncSubscription(sub)
 			require.NoError(t, err)
-			assert.Equal(t, len(jsBackend.subscriptions), tc.wantNatsSubsLen)
+			require.Equal(t, len(jsBackend.subscriptions), tc.wantNatsSubsLen)
 
 			if tc.givenManuallyDeleteSubscription {
 				// manually delete the subscription from map
@@ -1554,5 +1247,79 @@ func TestJetStream_NATSSubscriptionCount(t *testing.T) {
 			// empty the js.subscriptions map
 			require.NoError(t, jsBackend.DeleteSubscription(sub))
 		})
+	}
+}
+
+func defaultNatsConfig(url string) env.NatsConfig {
+	return env.NatsConfig{
+		URL:                     url,
+		MaxReconnects:           defaultMaxReconnects,
+		ReconnectWait:           3 * time.Second,
+		JSStreamName:            defaultStreamName,
+		JSStreamStorageType:     StorageTypeMemory,
+		JSStreamRetentionPolicy: RetentionPolicyInterest,
+	}
+}
+
+// getJetStreamClient creates a client with JetStream context, or fails the caller test.
+func getJetStreamClient(t *testing.T, serverURL string) *jetStreamClient {
+	conn, err := nats.Connect(serverURL)
+	if err != nil {
+		t.Error(err.Error())
+	}
+	jsCtx, err := conn.JetStream()
+	if err != nil {
+		conn.Close()
+		t.Error(err.Error())
+	}
+	return &jetStreamClient{
+		JetStreamContext: jsCtx,
+		natsConn:         conn,
+	}
+}
+
+func addJSCleanEventTypesToStatus(sub *eventingv1alpha2.Subscription, cleaner cleaner.Cleaner) error {
+	cleanEventType, err := getCleanEventTypes(sub, cleaner)
+	if err != nil {
+		return err
+	}
+	sub.Status.Types = cleanEventType
+	return nil
+}
+
+// TestEnvironment provides mocked resources for tests.
+type TestEnvironment struct {
+	jsBackend  *JetStream
+	logger     *logger.Logger
+	natsServer *server.Server
+	jsClient   *jetStreamClient
+	natsConfig env.NatsConfig
+	cleaner    cleaner.Cleaner
+	natsPort   int
+}
+
+// setupTestEnvironment is a TestEnvironment constructor.
+func setupTestEnvironment(t *testing.T) *TestEnvironment {
+	natsServer, natsPort, err := natstesting.StartNATSServer(evtesting.WithJetStreamEnabled())
+	require.NoError(t, err)
+	natsConfig := defaultNatsConfig(natsServer.ClientURL())
+	defaultLogger, err := logger.New(string(kymalogger.JSON), string(kymalogger.INFO))
+	require.NoError(t, err)
+
+	// init the metrics collector
+	metricsCollector := metrics.NewCollector()
+
+	jsClient := getJetStreamClient(t, natsConfig.URL)
+	jsCleaner := cleaner.NewJetStreamCleaner(defaultLogger)
+	jsBackend := NewJetStream(natsConfig, metricsCollector, jsCleaner, defaultLogger)
+
+	return &TestEnvironment{
+		jsBackend:  jsBackend,
+		logger:     defaultLogger,
+		natsServer: natsServer,
+		jsClient:   jsClient,
+		natsConfig: natsConfig,
+		cleaner:    jsCleaner,
+		natsPort:   natsPort,
 	}
 }
